enable mtimer interrupt

+++ src/timer.rs
@@ -1,7 +1,7 @@
 use core::ptr::{read_volatile, write_volatile};
 
 pub trait Oneshot {
-	fn start(&self, us: u64);
+	fn start(&self, us: usize);
 }
 
 pub struct MTimer {
@@ -21,10 +21,10 @@ impl MTimer {
 }
 
 impl Oneshot for MTimer {
-	fn start(&self, us: u64) {
+	fn start(&self, us: usize) {
 		unsafe {
-			let current_time = read_volatile(0x200BFF8 as *const u64);
-			write_volatile(0x2004000 as *mut u64, (current_time + us) as u64);
+			let current_time = read_volatile(0x200BFF8 as *const usize);
+			write_volatile(0x2004000 as *mut usize, (current_time + us) as usize);
 		};
 	}
 }
